Add context to CategoryContract.test.js
