+++ test/CategoryContract.test.js
@@ -1,4 +1,6 @@
 const CategoryContract = artifacts.require("CategoryContract");
+const ResearcherContract = artifacts.require("ResearcherContract");
+const UserContract = artifacts.require("UserContract");
 const SacToken = artifacts.require("SacToken");
 const IsaPool = artifacts.require("IsaPool");
 
@@ -8,7 +10,9 @@ contract("CategoryContract", (accounts) => {
   let instance;
   let sacToken;
   let isaPool;
-  let [msgSender, user1Address, user2Address] = accounts;
+  let userContract;
+  let researcherContract;
+  let [msgSender, user1Address, resea1Address] = accounts;
 
   const addCategory = async (name) => {
     await instance.addCategory(
@@ -19,6 +23,19 @@ contract("CategoryContract", (accounts) => {
       `${name} neutro`,
       `${name} parcialmente não sustentável`,
       `${name} totalmente não sustentável`
+    );
+  };
+
+  const addResearcher = async (name, address) => {
+    await researcherContract.addResearcher(
+      name,
+      "111.111.111-00",
+      "CPF",
+      "Brazil",
+      "SP",
+      "Jundiai",
+      "135465-005",
+      {from: address}
     );
   };
 
@@ -36,14 +53,45 @@ contract("CategoryContract", (accounts) => {
 
     await sacToken.addContractPool(isaPool.address, "0");
 
-    instance = await CategoryContract.new(isaPool.address);
+    userContract = await UserContract.new();
+    researcherContract = await ResearcherContract.new(userContract.address);
+
+    await userContract.newAllowedCaller(researcherContract.address);
+    await researcherContract.newAllowedUser(resea1Address);
+
+    instance = await CategoryContract.new(isaPool.address, researcherContract.address);
     await isaPool.newAllowedCaller(instance.address);
+
+    await addResearcher("Researcher A", resea1Address);
+
+  });
+
+  context("When will add a new category", () => {
+    context("When is not a researcher", () => {
+      it("should return error message", async () => {
+        const name = "Solo"  
+        await expectRevert(
+          addCategory(name, user1Address),
+          "Only allowed to researchers"
+        );
+      }) 
+    });
+    
+    context("When is a researcher", () => {
+      it("should create category", async () => {
+        const name = "Solo"  
+        await addCategory(name, resea1Address);
+        const categories = await instance.getCategories();
+    
+        assert.equal(categories[0].name, "Solo");
+      })  
+    });
   });
 
   context("When access category fields", () => {
     it("should have fields", async () => {
       const name = "Solo"
-      await addCategory(name);
+      await addCategory(name, resea1Address);
       const category = await instance.categories(1);
       
       assert.equal(category.id, 1);
@@ -60,14 +108,14 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("should create category", async () => {
-    await addCategory("Solo");
+    await addCategory("Solo", resea1Address);
     const categories = await instance.getCategories();
 
     assert.equal(categories[0].name, "Solo");
   });
 
   it("should add msg.sender in createdBy", async () => {
-    await addCategory("Solo");
+    await addCategory("Solo", resea1Address);
 
     const category = await instance.categories(1);
 
@@ -75,8 +123,8 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("should increment id of category when created", async () => {
-    await addCategory("Solo");
-    await addCategory("Solo 2");
+    await addCategory("Solo", resea1Address);
+    await addCategory("Solo 2", resea1Address);
 
     const categories = await instance.getCategories();
 
@@ -84,24 +132,24 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("should increment total of categories after a new is added", async () => {
-    await addCategory("Solo");
-    await addCategory("Solo 2");
+    await addCategory("Solo", resea1Address);
+    await addCategory("Solo 2", resea1Address);
     const categoryCounts = await instance.categoryCounts();
 
     assert.equal(categoryCounts, 2);
   });
 
   it("should return category list after when call getCategories", async () => {
-    await addCategory("Solo");
-    await addCategory("Solo2");
+    await addCategory("Solo", resea1Address);
+    await addCategory("Solo2", resea1Address);
     const categories = await instance.getCategories();
 
     assert.equal(categories.length, 2);
   });
 
   it("should be the same category in category and category list based on position", async () => {
-    await addCategory("Solo");
-    await addCategory("Solo 2");
+    await addCategory("Solo", resea1Address);
+    await addCategory("Solo 2", resea1Address);
     const categoriesList = await instance.getCategories();
     const category = await instance.categories(2);
 
@@ -109,7 +157,7 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("should create category with votes equal 0", async () => {
-    await addCategory("Solo");
+    await addCategory("Solo", resea1Address);
     const categories = await instance.getCategories();
 
     assert.equal(parseInt(categories[0].votesCount), 0);
@@ -120,7 +168,7 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("should vote and send tokens only to the passed id", async () => {
-    await addCategory("Solo");
+    await addCategory("Solo", resea1Address);
     await instance.vote(1, "100000000000000000000");
 
     const votes1 = await instance.votes(1);
@@ -131,7 +179,7 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("should increment 100 tokens when call vote function and send 100 tokens", async () => {
-    await addCategory("Solo");
+    await addCategory("Solo", resea1Address);
     await instance.vote(1, "100000000000000000000");
 
     const votes = await instance.votes(1);
@@ -140,7 +188,7 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("should have 150 tokens when call vote function and send 100 tokens and already have 50 tokens", async () => {
-    await addCategory("Solo");
+    await addCategory("Solo", resea1Address);
     await instance.vote(1, "50000000000000000000");
     await instance.vote(1, "100000000000000000000");
 
@@ -150,8 +198,8 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("should set amount of tokens that the user voted to category id when vote", async () => {
-    await addCategory("Solo 1");
-    await addCategory("Solo 2");
+    await addCategory("Solo 1", resea1Address);
+    await addCategory("Solo 2", resea1Address);
     await instance.vote(1, "100000000000000000000");
     await instance.vote(2, "50000000000000000000");
 
@@ -165,8 +213,8 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("should start with voted zero categories", async () => {
-    await addCategory("Solo 1");
-    await addCategory("Solo 2");
+    await addCategory("Solo 1", resea1Address);
+    await addCategory("Solo 2", resea1Address);
 
     const voted1 = await instance.voted(msgSender, 1);
     const voted2 = await instance.voted(msgSender, 0);
@@ -176,7 +224,7 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("should increment vote of users in same category", async () => {
-    await addCategory("Solo 1");
+    await addCategory("Solo 1", resea1Address);
 
     await transferTokensTo(user1Address, "50000000000000000000000");
     await transferTokensTo(user2Address, "50000000000000000000000");
@@ -190,7 +238,7 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("when more than one user voted in one category each user must have your tokens part", async () => {
-    await addCategory("Solo 1");
+    await addCategory("Solo 1", resea1Address);
 
     await transferTokensTo(user1Address, "50000000000000000000000");
     await transferTokensTo(user2Address, "50000000000000000000000");
@@ -206,17 +254,17 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("should return error message when try vote and dont has SAC Tokens", async () => {
-    await addCategory("Solo");
+    await addCategory("Solo", resea1Address);
     await expectRevert(instance.vote(1, 0, {from: user1Address}), "You don't have tokens to vote");
   });
 
   it("should return error message when try vote and has tokens but dont send any", async () => {
-    await addCategory("Solo");
+    await addCategory("Solo", resea1Address);
     await expectRevert(instance.vote(1, 0), "Send at least 1 SAC Token");
   });
 
   it("should subtract tokens from user after vote in a category", async () => {
-    await addCategory("Solo");
+    await addCategory("Solo", resea1Address);
     await transferTokensTo(user1Address, "50000000000000000000000");
     await instance.vote(1, "1000000000000000000000", {from: user1Address});
 
@@ -226,7 +274,7 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("should add tokens to isa pool after vote in a category", async () => {
-    await addCategory("Solo");
+    await addCategory("Solo", resea1Address);
     await transferTokensTo(user1Address, "50000000000000000000000");
     await instance.vote(1, "1000000000000000000000", {from: user1Address});
 
@@ -236,7 +284,7 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("should increment tokens in isa pool after many user votes", async () => {
-    await addCategory("Solo");
+    await addCategory("Solo", resea1Address);
     await transferTokensTo(user1Address, "50000000000000000000000");
     await transferTokensTo(user2Address, "50000000000000000000000");
     await instance.vote(1, "1000000000000000000000", {from: user1Address});
@@ -252,7 +300,7 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("should decrement votesCount in 1 when unvote with success", async () => {
-    await addCategory("Solo");
+    await addCategory("Solo", resea1Address);
     await transferTokensTo(user1Address, "50000000000000000000000");
 
     await instance.vote(1, "100000000000000000000", {from: user1Address});
@@ -265,7 +313,7 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("should return error message when user try unvote and dont voted yet", async () => {
-    await addCategory("Solo 1");
+    await addCategory("Solo 1", resea1Address);
 
     await instance.vote(1, "100000000000000000000");
 
@@ -276,7 +324,7 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("when a user unvoted with success should remove votes/tokens from voted mapping", async () => {
-    await addCategory("Solo 1");
+    await addCategory("Solo 1", resea1Address);
 
     await instance.vote(1, "100000000000000000000");
     await instance.unvote(1);
@@ -286,7 +334,7 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("when a user unvoted with success should remove votes/tokens from category", async () => {
-    await addCategory("Solo 1");
+    await addCategory("Solo 1", resea1Address);
 
     await instance.vote(1, "100000000000000000000");
     await instance.unvote(1);
@@ -296,8 +344,8 @@ contract("CategoryContract", (accounts) => {
   });
 
   it("when a user unvoted in a category your tokens should allowed back", async () => {
-    await addCategory("Solo 1");
-    await addCategory("Solo 2");
+    await addCategory("Solo 1", resea1Address);
+    await addCategory("Solo 2", resea1Address);
 
     await transferTokensTo(user1Address, "50000000000000000000000");
